Workaround for MySQL 8.4 and unknown users (#7136)
|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | N/A

#### Summary

See https://bugs.mysql.com/bug.php?id=114876 for details.
Failing test:
https://github.com/doctrine/dbal/actions/runs/17474044585/job/49629574307?pr=7134#step:7:98

The MySQL 8.4 branch has a known bug which causes it to complain about
the legacy password hashing plugin not being loaded if we try to login
with an unknown user. This _sometimes_ causes our tests to fail. With
this PR, I'm upcasting this error to the same exception which we would
get for invalid credentials.