Downgrade PHP for the DB2 workflow (#7134)
I have no idea why the DB2 workflow started to segfault. My best guess
is that there's an incompatibility with the latest patch release of PHP.
Let's see if pinning to an older release works.

Update: I've downgraded from PHP 8.4 to 8.3. That seems to "fix" it for
now and give us a green CI again. Let's monitor the issue and revert the
patch when the issues is solved.