Make options check strict again (#7141)
|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | N/A

#### Summary

While merging up the changes from #7132, I noticed that the new code is
in fact a little less strict than the old one. Previously, we required
the `PDO::PGSQL_ATTR_DISABLE_PREPARES` option to be set to `true` while
after the change we would accept any truthy value. I don't think that
it's a big deal, but I also believe that this change was not intended.
Thus, I'm restoring the old behavior.